requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
pandas==2.2.2
numpy==1.26.4
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
nextcord==2.6.0
colorama==0.4.6
discord==2.3.2
frozenlist==1.4.1
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
multidict==6.0.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
tqdm==4.66.4
typing_extensions==4.11.0
unicode==2.9
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4